Use a local build of webjars-play
Build 2.6.2 has a controllers.WebJarAssets which is
not compatible with Play and it is trying to be loaded
by Guice, but since the binary is not compatible, we
get an error.

After cloning webjars-play and running publishLocal
I was able to avoid this "broken" class.

The version used here is the last commit hash.